# Various performance improvements and cleanups to goforit