--- a/backend_test.go
+++ b/backend_test.go
@@ -28,16 +28,19 @@ func TestParseFlagsCSV(t *testing.T) {
 					"go.sun.money",
 					true,
 					[]RuleInfo{{&RateRule{Rate: 0}, RuleOn, RuleOff}},
+					nil,
 				},
 				{
 					"go.moon.mercury",
 					true,
 					[]RuleInfo{{&RateRule{Rate: 1}, RuleOn, RuleOff}},
+					nil,
 				},
 				{
 					"go.stars.money",
 					true,
 					[]RuleInfo{{&RateRule{Rate: 0.5}, RuleOn, RuleOff}},
+					nil,
 				},
 			},
 		},
@@ -80,6 +83,7 @@ func TestParseFlagsJSON(t *testing.T) {
 						{&MatchListRule{"host_name", []string{"apibox_789"}}, RuleOn, RuleContinue},
 						{&RateRule{0.01, []string{"cluster", "db"}}, RuleOn, RuleOff},
 					},
+					nil,
 				},
 				{
 					"go.sun.mercury",
@@ -87,6 +91,7 @@ func TestParseFlagsJSON(t *testing.T) {
 					[]RuleInfo{
 						{&RateRule{Rate: 0.5}, RuleOn, RuleOff},
 					},
+					nil,
 				},
 			},
 		},
@@ -112,10 +117,12 @@ func TestMultipleDefinitions(t *testing.T) {
 	const lastValue = 0.7
 
 	backend := BackendFromFile(filepath.Join("fixtures", "flags_multiple_definitions.csv"))
-	g, _ := testGoforit(0, backend)
+	g, _ := testGoforit(0, backend, enabledTickerInterval)
 	g.RefreshFlags(backend)
 
-	flag := g.flags[repeatedFlag]
-	assert.Equal(t, flag, Flag{repeatedFlag, true, []RuleInfo{{&RateRule{Rate: lastValue}, RuleOn, RuleOff}}})
+	flag, ok := g.flags[repeatedFlag]
+	assert.True(t, ok)
+	flag.enabledTicker = nil // we don't compare about comparing this
+	assert.Equal(t, flag, Flag{repeatedFlag, true, []RuleInfo{{&RateRule{Rate: lastValue}, RuleOn, RuleOff}}, nil})
 
 }