The primary improvement is in c7d09daa9cd49e1c03b43657869a7cb0c1cdb283.  Outside of that, here are benchmarks pre and post:

```
PRE:
BenchmarkEnabled50-4         2000000           651 ns/op         204 B/op           5 allocs/op
BenchmarkEnabled100-4        2000000           636 ns/op         204 B/op           5 allocs/op

POST:
BenchmarkEnabled50-4    	10000000	       161 ns/op	      48 B/op	       1 allocs/op
BenchmarkEnabled100-4   	20000000	        50.2 ns/op	       0 B/op	       0 allocs/op
```

You should read each commit individually; most are completely independent.

r? @vasi-stripe