@vasi-stripe 2018-04-16
Cool, I didn't know you could use tickers that way! Thanks for splitting this into logical commits, made it much easier to review.

LGTM